Added lstmconv2d-enc-dec model
To rebuild the model of the "Unsupervised Learning with LSTM", that uses MovingMNIST as well to be able to compare our results with this model. But how the state is copied (currently a bug in LSTMConv2D) and what kind of input is used for the decoder has still to be investigated, by looking more deeply into the code of the paper.